requirements.txt
apache-airflow~=2.2.4
apache-airflow-providers-postgres~=4.1.0
